move classes describing a hierarchy into a separate module
